requirements.txt
beautifulsoup4==4.5.3
bs4==0.0.1
PyYaml==3.12
six==1.10.0
requests==2.18.4
wheel==0.32.3
Pillow==5.0.0
